Simplify MouseService
